Optimize initial load and static homepage
