+++ middleware.ts
@@ -4,27 +4,6 @@ import { applySecurityHeaders, normalizeCookieOptions } from "@/lib/security";
 import { getSupabaseAnonKey, getSupabaseUrl } from "@/lib/supabase/keys";
 
 export async function middleware(request: NextRequest) {
-  // Handle auth callback redirects at root
-  if (request.nextUrl.pathname === "/") {
-    const code = request.nextUrl.searchParams.get("code");
-    const tokenHash = request.nextUrl.searchParams.get("token_hash");
-    const type = request.nextUrl.searchParams.get("type");
-
-    if (code) {
-      // Legacy ?code= parameter - redirect to /auth/callback
-      const url = request.nextUrl.clone();
-      url.pathname = "/auth/callback";
-      return NextResponse.redirect(url);
-    }
-
-    if (tokenHash && type) {
-      // New token_hash + type format - redirect to /auth/confirm
-      const url = request.nextUrl.clone();
-      url.pathname = "/auth/confirm";
-      return NextResponse.redirect(url);
-    }
-  }
-
   // /help is owned by root in some environments; serve the writable checklist page instead.
   if (request.nextUrl.pathname === "/help") {
     const url = request.nextUrl.clone();
@@ -33,7 +12,7 @@ export async function middleware(request: NextRequest) {
   }
 
   // Skip Supabase session refresh on public routes to reduce TTFB
-  const publicPaths = ["/", "/help", "/help-checklist", "/pricing", "/term", "/privacy", "/blog", "/coupang-csv"];
+  const publicPaths = ["/help", "/help-checklist", "/pricing", "/term", "/privacy", "/blog", "/coupang-csv"];
   const isPublicRoute = publicPaths.some(p => request.nextUrl.pathname === p);
   if (isPublicRoute) {
     const passthrough = NextResponse.next();
@@ -85,5 +64,5 @@ export async function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ["/((?!_next/static|_next/image|favicon.ico).*)"]
+  matcher: ["/((?!$|_next/static|_next/image|favicon.ico|manifest.json|robots.txt|sitemap.xml|sample\\.csv|sample_simple\\.csv).*)"]
 };
